monitoring/requirements.txt
psutil>=5.9
pyyaml>=6.0
requests>=2.31
